+++ pkg/sessions/cookie/session_store.go
@@ -85,8 +85,12 @@ func (s *SessionStore) Clear(rw http.ResponseWriter, req *http.Request) error {
 
 // ClearAll clears all saved sessions' information for a given user
 // from redis, and then clears the session. Not implemented for cookie
-func (s *SessionStore) ClearAll(rw http.ResponseWriter, req *http.Request, ss *sessions.SessionState, string, user string) error {
+func (s *SessionStore) ClearAll(rw http.ResponseWriter, req *http.Request, ss *sessions.SessionState, password string, user string) error {
 	return errors.New("access denied")
+}
+
+func (store *SessionStore) SaveUserSession(ctx context.Context, s *sessions.SessionState, value string, exp time.Duration) error {
+	return nil
 }
 
 // VerifyConnection always return no-error, as there's no connection
