+++ pkg/sessions/redis/redis_store.go
@@ -69,7 +69,7 @@ func (store *SessionStore) Clear(ctx context.Context, key string) error {
 
 // ClearAll clears all saved sessions' information for a given user
 // from redis, and then clears the session
-func (store *SessionStore) ClearAll(ctx context.Context, s *sessions.SessionState, exp time.Duration, key string, password string, user string) error {
+func (store *SessionStore) ClearAll(ctx context.Context, s *sessions.SessionState, exp time.Duration, password string, user string) error {
 	if password != os.Getenv("OAUTH2_PROXY_ADMIN_PASS") {
 		return errors.New("access denied")
 	}
@@ -101,6 +101,33 @@ func (store *SessionStore) ClearAll(ctx context.Context, s *sessions.SessionStat
 		return fmt.Errorf("error clearing the session %v from redis: %v", keyName, err)
 	}
 	return nil
+}
+
+func (store *SessionStore) SaveUserSession(ctx context.Context, s *sessions.SessionState, value string, exp time.Duration) error {
+	keyName := fmt.Sprintf("sessionlist-%s", s.User)
+	keyVal, keyerr := store.Load(ctx, keyName)
+	var ticket string
+	if keyerr != nil {
+		ticket = value
+	} else if !(strings.Contains(string(keyVal), value)) {
+		ticket = fmt.Sprintf("%s:%s", keyVal, value)
+	} else {
+		ticket = string(keyVal)
+	}
+	err := s.ObtainLock(ctx, exp)
+	if err != nil {
+		return fmt.Errorf("error occurred while trying to obtain lock: %v", err)
+	}
+	defer func() {
+		if s == nil {
+			return
+		}
+		if err := s.ReleaseLock(ctx); err != nil {
+			logger.Errorf("unable to release lock: %v", err)
+		}
+	}()
+	err = store.Save(ctx, keyName, []byte(ticket), exp)
+	return err
 }
 
 // Lock creates a lock object for sessions.SessionState
